Feature/add shopping cart

+++ src/main/java/org/example/springbootadventure/service/category/CategoryServiceImpl.java
@@ -4,10 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.example.springbootadventure.dto.category.CategoryDto;
 import org.example.springbootadventure.dto.category.CreateCategoryRequestDto;
 import org.example.springbootadventure.exceptions.EntityNotFoundException;
-import org.example.springbootadventure.mapper.BookMapper;
 import org.example.springbootadventure.mapper.CategoryMapper;
 import org.example.springbootadventure.model.Category;
-import org.example.springbootadventure.repository.book.BookRepository;
 import org.example.springbootadventure.repository.category.CategoryRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,42 +16,41 @@ import org.springframework.stereotype.Service;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
-    private final BookMapper bookMapper;
-    private final BookRepository bookRepository;
 
     @Override
     public Page<CategoryDto> findAll(Pageable pageable) {
         return categoryRepository.findAll(pageable)
                 .map(categoryMapper::toDto);
-
     }
 
     @Override
     public CategoryDto getById(Long id) {
         Category category = categoryRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Can't find category by id :" + id));
+                () -> new EntityNotFoundException("Can't find category by id: " + id)
+        );
         return categoryMapper.toDto(category);
-
     }
 
     @Override
     public CategoryDto save(CreateCategoryRequestDto categoryDto) {
-        Category entity = categoryMapper.toEntity(categoryDto);
-        return categoryMapper.toDto(categoryRepository.save(entity));
-    }
-
-    @Override
-    public CategoryDto update(Long id, CreateCategoryRequestDto categoryDto) {
-        Category category = categoryRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Can't find category by id :" + id));
-        categoryMapper.updateCategoryFromDto(categoryDto, category);
-
+        Category category = categoryMapper.toEntity(categoryDto);
         return categoryMapper.toDto(categoryRepository.save(category));
     }
 
     @Override
-    public void deleteById(Long id) {
-        categoryRepository.deleteById(id);
+    public CategoryDto update(Long id, CreateCategoryRequestDto categoryDto) {
+        Category existingCategory = categoryRepository.findById(id).orElseThrow(
+                () -> new EntityNotFoundException("Can't find category by id: " + id)
+        );
+        categoryMapper.updateCategoryFromDto(categoryDto, existingCategory);
+        return categoryMapper.toDto(categoryRepository.save(existingCategory));
+    }
 
+    @Override
+    public void deleteById(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find category by id: " + id);
+        }
+        categoryRepository.deleteById(id);
     }
 }
